api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.1
pinecone-client==3.0.0
google-generativeai==0.3.2
google-cloud-tasks==2.16.0
google-cloud-logging==3.9.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
python-multipart==0.0.6
structlog==24.1.0
prometheus-client==0.19.0
sentry-sdk==1.40.0
pytest==8.0.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
respx==0.20.2
ruff==0.1.14
mypy==1.8.0
